feat(article): add pagination dots
