+++ src/app/Scenes/Home/Components/ContentCards.tsx
@@ -1,11 +1,11 @@
 import { Box, Flex, Spacer, Text, Button, Touchable } from "@artsy/palette-mobile"
 import OpaqueImageView from "app/Components/OpaqueImageView/OpaqueImageView"
+import { PaginationDots } from "app/Components/PaginationDots"
 import { navigate } from "app/system/navigation/navigate"
 import { useScreenDimensions } from "app/utils/hooks"
 import { memo, useEffect, useRef, useState } from "react"
 import { FlatList, PixelRatio, ViewabilityConfig } from "react-native"
 import ReactAppboy from "react-native-appboy-sdk"
-import { PaginationDots } from "./PaginationDots"
 
 interface CardProps {
   item: ReactAppboy.CaptionedContentCard
